SMS, PSG: Mutexs, asserts, and corner-cases
Requests from the sound card come from a different thread to the main
emulator. This could result in the PSG code getting into a bad state
where the read-index gets ahead of the write-index.

Locking is added around psg_run_cycles, and the code to calculate the
minimum number of cycles needed for the sound card is moved there too.

The ring size is doubled, with the aim of keeping it only half-full. This
is attempted by discarding 20% of cycles when the ring is over half full.

Ring utilisation is added to the statistics menu.

A pair of asserts are added for hopefully-now-unreachable states to
prevent infinite loops.