sms: Select Mode4 when running without a bios
Some games will not start without some action from the VDP.

The previous patch removes assumptions about Mode4 being selected. Only
Mode4 is implemented, and an early return is used if another mode is
selected.

To have the VDP active from a boot without a BIOS, it needs to be left in
Mode4.