tms9928a: Rename from tms9918a
The tms9918a was a variant of the chip which supports composite video for
both input and output, and which has no support for component video.

The tms9928a (60 Hz) and tms9929a (50 Hz) are the chips that actually got
used in the consoles being emulated. These have no video input, and support
component video output.