sn76489: Constant channel output of +1 when tone register is 1
Earlier documentation used listed zero as the only tone register value to
output a constant +1 value.

Newer documentation on SMS Power! shows that both values of zero and one
will output a constant +1 value from a tone channel.

While we're at it, change the reset value of the lfsr to the MSB instead
of the LSB, as this is possibly closer to what hardware does.