maint: Introduce a run_mutex to prevent use after free
The console state is used in the main emulation loop. However, the state
can be freed from the GUI loop when closing or changing roms.

A new mutex is introduced to protect the run callback. This should be held
when freeing resources that the run_callback may use.

Going forward, this run_mutex may be able to replace the console-specific
state mutexes.