sms_vdp: Don't early return for unsupported video modes
Some games briefly pass through an unsupported video mode.
If we return early, the VDP state stops being updated, which can result
in a lockup. Eg, if the CPU is halted and waiting for an interrupt.

When an unsupported mode is configured, we now assume Mode 0. This way,
the line counter continues to increment and we continue to get interrupts.