sms_vdp: Increment the line counter before rendering the line
Previously, the line was rendered, and then afterward the line counter
was incremented to the value for the next line.

This caused a problem where the sprite overflow flag would appear to be
set on the wrong line number. As soon as the flag was set, the line number
would be incremented with no chance for the Z80 to run in between.

Now, the line number is incremented before rendering the line, allowing
the overflow flag and the line number to line up correctly.