tms9928a, sms_vdp: Apply register writes directly, remove regs_buffer
This change was added for a pair of games that no-longer need it.

It appears that buffering the register writes for the next line is the
wrong thing to do, as it can cause the line-interrupt to be off-by-one.

This will break save-state compatibility.

While we're editing the tms9928a code, move the point at which the line
counter is incremented to be consistent with the sms_vdp.