z80: Adjust z80 timing to run so long as we have >0 cycles available
Previously, we would build up a pool of cycles before running a single
instruction. Enough for any instruction followed by an interrupt.

Now, we run the z80 any time there is a positive number of cycles
available and keep track of the excess used.

This will hopefully simplify the timing of periodic events in the sms_vdp.