sms_vdp: Offset the h_counter to line up the frame interrupt with real hw
The z80 cycle-count % 228 is used as the time base for two different VDP
operations:

1. sms_vdp_run_one_scanline is called, which among other things will
   generate the frame interrupts. When this happens cycle_count % 228
   will be zero.

2. The h-counter, with the discontinuity in the count value occurring when
   cycle_count % 228 is zero.

However: The frame interrupt does not occur at the same time as the
h-counter discontinuity. Instead it happens part way through the count.

This patch adds an offset to the cycle_count used when calculating the
h-counter value. This offset brings the h-counter value at the point of
a frame-interrupt in line with what it is on real hardware.