maint: Locking around console initialization
This fixes an issue seen when using "Close ROM". The console would be set
to LOGO, and because running the logo screen doesn't require the RUNNING
state the callback would be called before it had been set.

This patch wraps a lock around everything from setting the console type
through to the initialization function returning.