uart: Functionality for sending sn76489 and ym2413 writes over UART
Note: This functionality is not hooked up in this patch.
      Register writes need to be redirected to uart_write_sn76489
      and uart_write_ym2413. uart_open also needs to be called with
      a valid path, such as "/dev/ttyUSB0".

This patch provides a handful of functions that allow the sound generation
to be sent to real sound chips hooked up to a micro-controller.

For each register write, two bytes are sent at 28800 baud. A separate
thread is used for writing to the UART, to prevent delays to the main
emulation thread.