ym2413: Lookup functions for the log-sin and exp tables
Note that a typedef is used to indicate the sign-magnitude integers that
these tables use.

Sign-magnitude is needed for the log-sine table, otherwise, the peak and
trough values would be identical (+0 and -0)

The output of the exp table also uses sign-magnitude, as +0 and -0 have
different transistors in the YM2413 DAC so may represent two different
output levels.