ym2413: Separate damp-attack transition for modulator and carrier
Comparing to actual hardware, the modulator does not transition to attack
at the same time as the carrier. It has its own independent damp-attack
transition.

This can be observed using attack-rate 15 and decay rate 0 for both the
modulator and carrier. As the carrier gets to release, its attack can
begin instantly. The modulator is still at eg-level 0 from the previous
note played, so its damp can be seen in the waveform before it transitions
to attack and jumps back to eg-level 0.