ym2413: Only run the modulator's envelope generator during key-down
Experimenting with a real chip, it appears that the modulator's envelope
generator doesn't run while the key is up. This can be observed by the
presence of the modulator's damp-before-attack at the start of a note.

If the envelope generator were running, the release of the prevous note
would have, given enough time,  bought the eg-level high enough that the
damp wouldn't be noticed.