ym2413: Save state support
Implementations for importing/exporting YM2413 state.

Note that the audio-control register is not directly included. This is a
Master System register, but adding it to the Master System state would
break compatibility with the current save format. For now, it's value is
assumed based on the presence or absence of a YM2413 state.

Eventually the plan is to make the save format more flexible so that new
values can be added to existing chips.