sn76489: Fixup noise channel volume
During the initial development of the sn76489 support, the implementation
was based on a document that suggested the noise channel had only half
the level swing of the tone channels.

This is found to not be the case, the swing of the noise channels is
measured to be about on par with the tone channels. The measured level
was actually ~90% compared to the tone I tested, but this can be
explained by low-pass filtering within the console.

While we're at it, change the initial values of all outputs to '1'.
Nothing was gained by having a mix of '1' and '-1'.