sn76489, ym2413: Simplify overrun protection and add it to the ym2413
The sn76489 implementaiton contained a 'rubber band' sort of overrun
protection, where it would subtract cycles if the buffer was starting to
get full. While this worked, it did mean introducing some timing variance
to the waveform.

Now, the mechanism is simplified: If the samples we're about to generate
would overrun the ring buffer, then discard the oldest quarter of the
buffer.

This new method is applied both to the sn76489 and the ym2413.