sn76489, ym2413: Remove assumption about power-of-two sized requests
Previously, when coping samples out of the ring buffer, an assumption was
made that the request would come in a power-of-two size. This assumption
meant a single request would never need to handle crossing the boundary
at the end of the ring buffer.

By adjusting the code to not rely on a linear read from a start position,
we can handle arbitrarily sized reads. (so long as they're smaller than
the buffer size itself).