ym2413: Lower the volume of the YM2413
The volume of the YM2413 was previously chosen to give a balance between
the two sound chips that sounded similar to recordings of the Japanese
Master System. This gave a per-melody-channel level of ±6126.

However, that was done before the SN76489 had been given a logarithmic
volume implementation. Now, that chosen volume level appears too loud
compared to the SN76489.

This patch reduces the level to ±4096 per channel. This both brings the
level to a better match with the PSG, and should also help to reduce the
chance of clipping the int16_t samples.