tms9928a: Add collision detection, improve 5th-sprite behaviour
Testing was done on an SC-3000H to find the behaviour of real hardware.

This patch brings Snepulator closer to that behaviour, although does not
yet implement the behaviour when the 5th-sprite bit is unset. On the
SC-3000H, this results in the 5th-sprite-number field reporting the number
of sprites. Possibly, in silicon this constantly updates as the sprite
list is traversed, and only stops updating once a 5th sprite is hit in
a line.

Note: The removed line-interrupt flag is for the SMS VDP. There is no
      line-interrupt on the tms9928a.