vgm: Initial VGM player implementation
This is the initial implementation of a VGM player for Snepulator, able
play music for the sn76489 and ym2413.

Code is in place to ignore writes to a couple of MSX sound chips, as some
VGM files begin with initialization of these, even though the music is
written only to the ym2413.

Still to do: Some kind of visualisation.