windows: Initial Snepulator build for Windows
This patch updates the build script to add a "windows" option, to build
a  64-bit windows executable with mingw. As no code is in place pet for
Windows directory reading, the 'Open' dialogue is disabled for the Windows
build.

Initial testing with Wine shows the following issues:
 - Environment Error: ${HOME} not defined.
 - Sound is clipping. Badly.
 - White lines in logo are transparent
 - Save states seem corrupted

The following .dll files are currently required:
 - SDL2.dll
 - libssp-0.dll
 - libpng16-16.dll
 - libwinpthread-1.dll
 - zlib1.dll

Note: If we switch to libspng, then we could remove libssp and libpng from
      the list of required .dll files.