build: Automatically do a clean build when switching compilers
At build time, a build_id.txt file is created to keep track of the
compiler and imgui version used for the most recent build. When one
of these changes (eg, doing a Windows build after a Linux one),
automatically do a clean build so that we don't mix the different types
of object files.