windows: Initial support for Open dialogue
This allows traversing the directories in the current drive, which will
be the drive of the user's profile. For now, even on windows the '/' is
used as the directory separator within Snepulator.

Windows doesn't give us `d_name` in the `struct dirent`, so instead we
switch to using `stat` to check if an entry is a directory.

Changing to a different drive letter is not yet implemented.