gui: Hide "../" in the root directory for Windows too
Some windows systems report a ".." entry in the root directory, while
others don't. We already hide this entry in linux by checking if the
path is exactly "/"

This patch updates the check so that instead of matching on exactly
a unix-style "/", we instead check that there is only one '/' in
the path.