sg-1000: Default to interrupt-mode 1
Normally, the z80 would power up in interrupt-mode 0, which has not been
implemented. However, nothing on the SG-1000 actually uses mode-0. Games
normally switch to interrupt-mode 1 early on when they are run.

However - Some SG-1000 homebrew may only have been tested on a Master
System or flash cart. These have a BIOS or menu that configures the
interrupt-mode before the game is started.

To enable compatibility with SG-1000 homebrew that doesn't set interrupt
mode 1 explicitly, have it be the default.