maint: Don't move to RUN_STATE_INIT from RUN_STATE_EXIT
During snepulator_reset and snepulator_error, state.run is set to
RUN_STATE_INIT to stop the emulation. However, if we were trying to exit
the application, this would overwrite RUN_STATE_EXIT, and the emulation
thread may never exit, leaving the main thread hanging at pthread_join.

This patch adds checks, so that the affected functions will only move to
RUN_STATE_INIT if the state was RUNNING or PAUSED.