timing: Use clock_rate variable instead of a get_clock_rate () function
No reason to call a function repeatedly when the result only ever changes
when switching between NTSC and PAL.

While we're at it, a common definition is added for the clock rates:
 - NTSC colourburst frequency
 - 4/5 PAL colourburst frequency.

Note, this is potentially 1-2 ppm faster than the real crystals used on
the master system, as they use a 53.6931MHz or 53.2034MHz crystal and
divide it by 15. The difference however should be negligible.