sn76489, ym2413: Nonzero initial vaule for context clock_rate
The first time sn76489_run_cycles / ym2413_run_cycles is called from
the console emulation code, it gets the console's clock rate.

However, if the first caller was actually a request for more samples for
the sound card, then it would be called with a clock rate of 0 Hz,
resulting in a divide by zero.

To avoid this, just assume an NTSC clock initially. Until the first call
for the console it will only be outputting silence anyway.

While touching the files, change shortfall to be 32-bit. It shouldn't be
very big.