video: Allow frames to be queued to avoid skipped and duplicated frames
Emulation is currently ticked every 1ms. However, this is not a clean
divider of the frame rate. Some frames will be emitted 16 ms apart while
others are emitted 17 ms apart. This results in some frames being shown
twice, while others are skipped entirely.

To reduce this stutter in the video, don't just show the most recent
frame. Instead, queue the frames as they are generated. When we need to
show a frame on the screen, pick the next frame in the queue.

To avoid adding too much latency, the queue is limited to the next frame
to be shown, and the frame that comes after that. If a third frame is
generated, it will replace the frame at the end of the queue.

Locking is adjusted slightly, as when reading we only really need to
lock while modifying the read index.

A future goal for this is to reduce the use of memcpy farther by passing
around frame pointers. These pointers could contain additional data such
as the frame's active area size, which currently may be updating slightly
earlier than the buffer.