video: Separate video filters into separate fragment shaders
Previously, all four video options were implemented in a single fragment
shader. The type of shading was passed as a uniform, for a switch
statement within the shader.

This patch splits the four options into separate fragment shaders. This
reduces the amount of branching within the shader.

All shaders are compiled at startup, and the selected shader is used when
drawing the frame.

We can also remove video_show_border and video_has_border:
 - video_has_border was mostly for the Game Gear, where we already black
   out areas in the frame buffer outside the active area.
 - video_show_border was tied to specific shaders, enabled for all except
   the dot-matrix. This can just be baked into the shaders.

As the enum is now used to select the shader program, rename it from
Video_Filter to Shader.