video: Non-integer scaling for dot-matrix shader
The get_pixel function us updated, so that instead of getting a pixel
in the output resolution, it is instead getting a pixel from a virtual
texture that has been integer scaled and had the grid lines applied.

Linear interpolation is used to stretch the last bit out. To keep the
weight of the lines looking consistent, the colour values are converted
to a linear colourspace before interpolation.