video: Non-integer scaling for scanlines shader
For thy X axis, we continue to interpolate with smoothstep.
This seems to do a reasonable job of preventing shimmer.

For the Y axis, the Y position within the line is used to darken the
transition between lines, using a simple triangular curve. Note that
underneath the darkening, the Y asis  is a nearest-neighbour scale with
no interpolation, so on non-integer scales some lines will be a pixel
thicker than others. However, this is hopefully hidden by the darkened
areas.

This looks decent for integer scales, but more work is needed for
non-integer scales to get the brightness more uniform, especiall at
smaller sizes.