video, sms: Consider left column blanking to be a lower output resolution
Rather than having special handling just for left column blanking, treat
it as lowering the output resolution.

When  left column blanking is enabled, the image will now be centred
rather that appearing closer to the right edge of the screen.

This also means we can now remove the 'options' uniform for the shaders.