maint: Move mutexes into snepulator state struct and change include use
This was started as an attempt to use native Windows lock and unlock
functions instead of pthreads. However, it looks like libgcc uses
pthreads internally, so removing the dependency is probably more pain
than it is worth.

Code changes done during this experiment however do leave us in a better
place to try again in the future, and make things just a hair tidier. So
rather than throwing everything out, deliver some of the changes:

 1. Instead of using PTHREAD_MUTEX_INITIALIZER with global mutexes, move
    the mutexes into the Snepulator_State struct and initialise with
    pthread_mutex_init.

 2. Rather than including pthread.h in every file, include it from within
    snepulator.h

 3. In line with adding pthread.h, also include sstdbool.h, stdint.h,
    snepulator_compat.h, and snepulator_types.h from snepulator.h. This
    means that snepulator.h now automatically includes the headers for
    all the types that it uses. Any file that includes snepulator.h will
    not need to add extra includes just to get it building.

 4. Slight adjustment to snepulator_compat.h, now including windows.h
    when building for Windows, so that other files don't need to check
    if they need to include it themselves.