ym2413: Fixup clock_rate initialization
The ym2413 clock rate is updated when the emulation thread gets called.
However, ym2413_init needs to initialize context->clock_rate with a
sensible value in case the audio_callback runs before the first emulation
callback.

While code for this was in place, a few lines farther down the elements
of YM2413_Context were being initialized to zero. This was likely left
over from before these values were stored in the context struct.