m68k: First instructions (move.b, tst.w, tst.l, lea, movem.w, movem.l, bra, bne)
For now, the strategy is to have a 64K-entry array for the 16-bit
opcodes. Each entry in the array is populated with a function pointer
when implemented.

The first handful of instructions are implemented, although without
timing information yet. This is covers the first eight instructions run
by Sonic the Hedgehog.

 - move.b Dn ← d(An)
 - tst.w (xxx).l
 - tst.l (xxx).l
 - lea d(PC)
 - movem.w (An)+  (memory-to-register)
 - movem.l (An)+  (memory-to-register)
 - bra
 - bne