m68k: Check state.run before each instruction (debug)
This checks that we're in RUN_STATE_RUNNING before each instruction is
executed. This stops us from continuing as soon as something unimplemented
is hit, rather than after the line finishes.

Ideally, the CPU implementation shouldn't be looking at Snepulator_State,
so this should eventually be reverted or reworked to use something more
specific to this particular instance of the m68k.