sms, colecovision: Allow hard-reset when running BIOS
Previously, with no cartridge was loaded, a hard-reset would return to
the logo screen instead of restarting with the BIOS.

snepulator_system_init is now updated to take in a console, so that the
current console can be passed in when restarting. Other callers should
get the console from snepulator_select_console_for_rom.