vgm: Error for unsupported data-block
If opening a Mega Drive VGM, there will likely be data blocks.

Previously, this would result in a lock-up as the data-block would be hit
repeatedly without ever reaching a delay command.

Now, hitting a data-block is considered an error, and the loop can be
broken by leaving the RUNNING state.

While we're at it, treat any unknown command as an error to prevent other
possible lockups when loading VGMs which use unsupported chips.