maint: Separate run-state for errors
Previously, if an error occurred during initialization, but a ROM was
given as an argument; because the error had left us in RUN_STATE_INIT,
the ROM would be loaded and start to run behind the error message.

Now, when we hit an error, a new RUN_STATE_ERROR is used.

snepulator_set_rom will not load a ROM if we are currently in the error
state. The error-state is dismissed once the error window is closed.

Upon closing the error window, the ROM is closed and we return to showing
the logo screen.