z80: Correct R register timing
Initially, Snepulator incremented the Z80 R register only once per
instruction. However, an edge case has been found where this results
in a game lockup.

In Final Bubble Bobble, a loop reads the R register looking for a
non-zero when masked with 0x07. This loop is eight instructions long,
so the three least significant bits of R do not change from one loop
to the next.

The R register is updated to follow the behavior of real Z80 hardware,
based on "The Undocumented Z80 Documented" by Sean Young.

 - Continue to increment R during HALT, as the HALT behaviour is
   actually to repeatedly run the HALT instruction.

 - Increment R for the cb, dd, ed, and fd prefixes

 - Increment R when accepting an interrupt