sms: Fixups for various Korean games
32 KB games are a bit of a spercial case, as they normally wouldn't need
a mapper. At a previous point in time, that's what Snepulator did. If a
game was only 32 KiB, the assumption was SMS_MAPPER_NONE.

However, that assumption breaks some homebrew, which can be both small,
but may also expect a Sega mapper for enabling SRAM saves. So, the
assumption of SMS_MAPPER_NONE for 32 KiB ROMs was removed. Instead, we
use whichever mapper the game tries to make use of.

Unfortunately, a number of Korean games happen to write to addresses used
by mappers despite not actually having mappers.

To work around this problem, hints are added for the Korean games which
run into this problem.