sms: Hint to disable the memory_control register
Some Korean games make dodgy writes to the memory control register which
disable the cartridge slot.

Previously, one game was fixed by preventing writes that would disable
everything (cart, bios, and ram). However, Power Boggle Boggle does a
write which disables the cart but not everything.

As a work around, a hint is added to disable the memory control register
on a case-by-case basis for games that misuse it.