midi: First sounds :D
Available ym2413 channels are stored in a ring buffer, so that the
oldest-free channel will be the next to be used. This way, we hopefully
give some time for the decay to run.

Listening to a couple of files, it's clear that a few features are
missing, like correct tempo, volume, and percussion.