midi: Velocity support
For now, this is mapped to the ym2413 volume with the following formula:

 (max velocity * max volume) - (velocity * volume)

The least-significant ten bits are discarded, keeping only the top four
bits of the 14-bit result.

Possibly the velocity-to-volume behaviour should have some kind of curve,
but it doesn't appear super consistent. As the instruments are fixed in
ROM, we don't get to change the timbre based on velocity.