midi: Introduce RUN_STATE_STOP to return to INIT when finished playing
Switching back to INIT and calling snepulator_rom_set (NULL) cannot be
done directly from the run callback. This is because the run_mutex is
already held, but snepulator_reset will wait for the lock to be free.

Instead, RUN_STATE_STOP is used to indicate the playback is stopped,
and the transition to INIT can occur once the lock has been released.