video: Separate active area and backdrop into separate textures
This patch separates the backdrop colour from the active area of the
video output texture. This is a step towards simplifying the rendering
code and making the output resolution more flexible.

For this initial step, the left and right borders are removed from the
video-out texture, which is renamed to active area. The active area
texture still has a fixed size however, so contains an unused portion
above and below the actual active area. The eventual goal will be to
remove any use of start_x and start_y, including exactly and only the
active area in the active area texture.

For the backdrop texture, an initial attempt was made to use a 1D
texture. But this resulted in the shader not running, it would show
through to the "should not be visible" warning message behind the video
out area. A one-pixel-wide 2D texture also did not appear to work
correctly, with only pixel (0, 0) showing up correctly. Through trial
and error, it was found that a one-pixel-high texture works correctly.