video: Draw only what will be displayed in frame->active_area
Previously active_area included various borders. Top and bottom for home
consoles, and borders on all sides for the Game Gear.

With this change, the final size of the active_area buffer is decided
when drawing the first line of the frame, and only the displayed content
is rendered to it.

Some awkward math is used to deal with the cropping for Game Gear mode.
This should be revisited, as checking the bounds of every pixel is kinda
inefficient.