video: Option to disable the backdrop border
Some games, like the Moon Patrol prototype on the ColecoVision, don't
do a good job of keeping colour transitions in the border to a specific
line, leading to a flickering effect.

To prevent this from becoming distracting, add an option to prevent the
border from being shown. This is implemented by passing an all-black
backdrop texture to the shader.