tms9928a: Text Mode implementation
Text mode gives a 40x24 grid of characters, each character being 6x8
pixels in size. This allows more text to be shown on screen, despite
a lower overall resolution of 240x192.

Calls from sms_vdp are also hooked up for Multicolour mode, which was
missed when Multicolour mode was implemented for the tms9928a.