sms: Initial mouse support for paddle games
When using paddle input, allow the player to use the mouse instead of
gamepad to move the paddle.

When clicking into the window with the controller type as paddle, the
mouse is captured by the window. Relative position is added to the paddle
position with a (for now) hard-coded scaling factor of 0.25. Internal
representation of the paddle position is now floating-point, rounded to
decimal when the Master System goes to read the value.

The mouse-capture can be exited with escape or right-control.

The 'RANGE' macro is also renamed to 'CLAMP', as this gives a slightly
better description of what it's doing.