sms: Initial Sports Pad support
This patch implements the Relative mode of the Sports Pad controller,
which is used by Great Ice Hockey and Sports Pad Football.

In this mode, the game holds TH normally-high. To make a reading of the
change in trackball position, the game toggles TH low-high-low-high.

After each TH transition, plus a 40-80 µs delay, a nibble is sampled from
the RIGHT, LEFT, UP, DOWN bits:

 - Nibble 1: X (most-significant bits)
 - Nibble 2: X (least-significant bits)
 - Nibble 3: Y (most-significant bits)
 - Nibble 4: Y (least-significant bits)

TL and TR are used for the two buttons on the Sports Pad.

The following changes are made to get this working:

 - In the mouse state in snepulator.h, cursor_button is split into
   mouse_button_left and mouse_button_right to allow two-button input.

 - Mouse input is captured using SDL_SetRelativeMouseMode. This hides
   the PC's mouse cursor and keeps its input going to snepulator. To
   release the mouse, press <Escape> or <Right-Ctrl>.

 - Following a (fixed for now) scaling factor. the relative mouse motion
   is accumulated in trackball_delta until the console requests a reading.
   The scaling factor, 0.04, was chosen to give my Kensington SlimBlade
   trackball a similar response in-game to the real Sports Pad.

 - Sports Pad added to gamepad type enum and menubar.

 - Timing is based on the number of cycles that the z80 has run at the
   point of IO writes. The cycle count is also passed when reading
   nibbles, which will be needed in the future for absolute-mode.

 -