sms: Sports Pad absolute-mode
The Sports Pad, with the switch set to 'Sports', actually has two modes.

 - The relative mode, implemented in the previous patch, which relies
   on the console generating strobe pulses on the TH line to clock
   the data. The clocked data is split over four nibbles, which contain
   X and Y relative coordinates.

 - The absolute mode, where the data is self-clocked by the Sports Pad
   using the TH / TL lines to clock the data. The clocked data is split
   over five nibbles, the extra nibble containing the two buttons.

The previous patch implemented the relative mode. This patch adds the
absolute mode.

Note that on a real Sports Pad, the choice of mode appears to be made
shortly after power-on (or, shortly after switching from 'Control' to
'Sports'). Once a mode is selected, it will not automatically change
to the other mode without un-plugging/re-plugging or toggling the switch.

On this emulated Sports Pad, the behaviour is slightly different. The
cable can't easily be un-plugged or the switch toggled, so instead the
mode is selected dynamically. If there has been no activity on the TH
line for 3 seconds, run in absolute-mode. Otherwise, run in relative mode.

As the Sports Pad implementation now does different things with TL / TR,
gamepad_trackball_get_nibble is renamed to gamepad_trackball_get_port,
to indicate it's touching more than just the low four bits.