sms, gui: Interface to adjust trackball and paddle sensitivity
The trackball is adjusted in units of fractional trackball-pixels per
host-pixel. The default value of 0.04 gives a similar response to the
real thing when using a trackball on my PC; of course this will depend
on the host mouse-speed setting.

The paddle is adjusted in units of fractional paddle-steps per host-pixel.
Each paddle-step being one step through the 256 possibl positions that
the paddle can report through its full rotation.
The default value of 0.25 makes Woody Pop fairly playable with a mouse.

These new settings are added to the input configuration dialogue, using
a tabbed interface to choose the type of emulated device being configured.